@@ -24,20 +24,20 @@
     "@apollo/server": "^4.11.2",
     "@arcath/utils": "^3.15.0",
     "@epic-web/totp": "^2.0.0",
-    "@prisma/client": "^5.21.1",
-    "@remix-run/node": "^2.13.1",
-    "@remix-run/react": "^2.13.1",
-    "@remix-run/serve": "^2.13.1",
-    "@tanstack/react-query": "^5.59.16",
+    "@prisma/client": "^6.0.1",
+    "@remix-run/node": "^2.15.1",
+    "@remix-run/react": "^2.15.1",
+    "@remix-run/serve": "^2.15.1",
+    "@tanstack/react-query": "^5.62.8",
     "adm-zip": "^0.5.16",
     "bcrypt": "^5.1.1",
-    "bullmq": "^5.21.2",
+    "bullmq": "^5.34.2",
     "cancant": "^1.1.0",
     "date-fns": "^4.1.0",
-    "framer-motion": "^11.11.10",
+    "framer-motion": "^11.15.0",
     "graphql": "^16.9.0",
     "is-ip": "^5.0.1",
-    "isbot": "^5.1.17",
+    "isbot": "^5.1.18",
     "mdx-bundler": "^10.0.3",
     "mkdirp": "^3.0.1",
     "node-cron": "^3.0.3",
@@ -47,48 +47,48 @@
     "qrcode": "^1.5.4",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-hotkeys-hook": "^4.5.1",
-    "react-use": "^17.5.1",
+    "react-hotkeys-hook": "^4.6.1",
+    "react-use": "^17.6.0",
     "rehype-autolink-headings": "^7.1.0",
     "rehype-slug": "^6.0.0",
     "remark-gfm": "^4.0.0",
     "tsx": "^4.19.2",
-    "uuid": "^11.0.2"
+    "uuid": "^11.0.3"
   },
   "devDependencies": {
-    "@eslint/compat": "^1.2.2",
-    "@eslint/eslintrc": "^3.1.0",
-    "@eslint/js": "^9.13.0",
-    "@faker-js/faker": "^9.1.0",
-    "@remix-run/dev": "^2.13.1",
+    "@eslint/compat": "^1.2.4",
+    "@eslint/eslintrc": "^3.2.0",
+    "@eslint/js": "^9.17.0",
+    "@faker-js/faker": "^9.3.0",
+    "@remix-run/dev": "^2.15.1",
     "@types/bcrypt": "^5.0.2",
     "@types/nprogress": "^0.2.3",
     "@types/papaparse": "^5.3.15",
     "@types/qrcode": "^1.5.5",
-    "@types/react": "^18.3.12",
-    "@types/react-dom": "^18.3.1",
+    "@types/react": "^19.0.1",
+    "@types/react-dom": "^19.0.2",
     "@types/uuid": "^10.0.0",
-    "@typescript-eslint/eslint-plugin": "^8.12.2",
-    "@vitest/coverage-istanbul": "^2.1.4",
-    "@vitest/eslint-plugin": "^1.1.7",
+    "@typescript-eslint/eslint-plugin": "^8.18.1",
+    "@vitest/coverage-istanbul": "^2.1.8",
+    "@vitest/eslint-plugin": "^1.1.17",
     "autoprefixer": "^10.4.20",
     "env-cmd": "^10.1.0",
-    "eslint": "^9.13.0",
-    "eslint-import-resolver-typescript": "^3.6.3",
+    "eslint": "^9.17.0",
+    "eslint-import-resolver-typescript": "^3.7.0",
     "eslint-plugin-import": "^2.31.0",
-    "eslint-plugin-import-x": "^4.4.0",
+    "eslint-plugin-import-x": "^4.5.1",
     "eslint-plugin-jsx-a11y": "^6.10.2",
     "eslint-plugin-react": "^7.37.2",
-    "eslint-plugin-react-hooks": "^5.0.0",
-    "globals": "^15.11.0",
-    "postcss": "^8.4.47",
-    "prettier": "^3.3.3",
-    "prisma": "^5.21.1",
-    "tailwindcss": "^3.4.14",
-    "typescript": "^5.6.3",
-    "typescript-eslint": "^8.12.2",
-    "vite-tsconfig-paths": "^5.0.1",
-    "vitest": "^2.1.4"
+    "eslint-plugin-react-hooks": "^5.1.0",
+    "globals": "^15.13.0",
+    "postcss": "^8.4.49",
+    "prettier": "^3.4.2",
+    "prisma": "^6.0.1",
+    "tailwindcss": "^3.4.16",
+    "typescript": "^5.7.2",
+    "typescript-eslint": "^8.18.1",
+    "vite-tsconfig-paths": "^5.1.4",
+    "vitest": "^2.1.8"
   },
   "engines": {
     "node": ">=18.0.0"
